pyproject.toml
[project]
name = "officeclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nanobot-ai",
]

[tool.uv.sources]
nanobot-ai = { git = "https://github.com/vvbeliaev/nanobot", rev = "officeclaw" }
